DataGridViewNumericUpDownEditingControl
