+++ Source/Frontend/UI/UI.csproj
@@ -431,6 +431,7 @@
     <Compile Include="UIConnector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Components\DataGridViewDraggable.cs" />
+    <Compile Include="Components\DataGridViewNumericUpDownEditingControl.cs" />
     <Compile Include="UI_Extensions.cs" />
     <Compile Include="UI_VanguardImplementation.cs" />
     <EmbeddedResource Include="Components\Blast Editor\RTC_ColumnSelector_Form.resx">
